refactoring for new linter-rules
